requirements.txt
# Core Dependencies
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Apache Airflow (compatible with Python 3.12)
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.10.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Additional Utilities
python-dateutil==2.8.2

# Web Framework (FastAPI)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
